fix(docs): guard every highlighter entry point and enforce the server boundary
- Add a calibrated marker per Shiki entry point the docs import: `scopeName`
  for `@shikijs/langs` and `Invalid recursionLimit; use 2-20` for the JS regex
  engine. Minified grammar-only (171KB) and engine-only (57KB) bundles both
  scanned clean under the previous core-only marker set.
- Flatten the nested `createRequire` and drop `JSON.stringify` so calibration
  resolves in one hop and matches the chunk scan byte-for-byte.
- Pin `@shikijs/langs` to `4.4.3`; `shiki` pins it exactly, so a floating caret
  could install a second grammar copy.
- Split `highlightSample` into `_lib/code-sample.ts` and put `server-only` on
  the component, making the server boundary enforced rather than asserted.
  Vitest aliases `server-only` to its `react-server` no-op, since jsdom does
  not resolve that export condition.
- Move the `CodeSample` tests beside their module, drop the `render` alias, and
  scope the hex assertion to style-attribute values.
- Trim rationale comments back to the non-obvious logic.

Closes #675
