+++ apps/docs/app/_lib/code-block.ts
@@ -1,7 +1,6 @@
 /**
  * Shared surface for every docs code block — the MDX `pre` and the
- * hand-written samples in `_pages`. Shiki emits the `pre` as a tab stop in both
- * paths, so the focus ring travels with the surface.
+ * hand-written samples in `_pages`.
  */
 export const CODE_BLOCK_SURFACE =
   'nx:typography-code-block nx:bg-container nx:border nx:border-border-default nx:rounded-md nx:p-4 nx:overflow-x-auto nx:focus-visible:outline-2 nx:focus-visible:outline-focus-default nx:focus-visible:outline-offset-(--focus-offset)';
