+++ apps/docs/package.json
@@ -19,13 +19,14 @@
     "@nexus_ds/core": "workspace:*",
     "@nexus_ds/react": "workspace:*",
     "@nexus_ds/tailwind": "workspace:*",
-    "@shikijs/langs": "^4.4.3",
+    "@shikijs/langs": "4.4.3",
     "next": "^15.1.6",
     "react": "^19.2.3",
     "react-dom": "^19.2.3",
     "rehype-pretty-code": "^0.14.5",
     "rehype-slug": "^6.0.0",
     "remark-gfm": "^4.0.1",
+    "server-only": "^0.0.1",
     "shiki": "^4.4.3",
     "sonner": "^2.0.7",
     "tw-animate-css": "^1.4.0"
